精度重构，feature之前精度校正完成

+++ src/main/java/com/westlake/air/pecs/domain/bean/score/ExperimentFeature.java
@@ -11,18 +11,18 @@ import java.util.List;
 @Data
 public class ExperimentFeature {
 
-    float rt;
+    double rt;
 
-    float intensity;
+    double intensity;
 
-    float peakApexInt;
+    double peakApexInt;
 
-    float bestLeft;
+    double bestLeft;
 
-    float bestRight;
+    double bestRight;
 
-    float totalXic;
+    double totalXic;
 
-    List<Float> hullRt;
-    List<Float> hullInt;
+    List<Double> hullRt;
+    List<Double> hullInt;
 }
